Make connectWithBackoff its own module
