+++ lib/connect.js
@@ -3,7 +3,6 @@ var parallel = require('run-parallel');
 var series = require('run-series');
 var clone = require('lodash.clone');
 var MongoClient = require('mongodb').MongoClient;
-var backoff = require('backoff');
 var Connection = require('./model');
 var parseURL = require('mongodb/lib/url_parser');
 var debug = require('debug')('mongodb-connection-model:connect');
@@ -77,36 +76,6 @@ function validateURL(model, done) {
   }
 }
 
-function connectWithBackoff(model, done) {
-  if (!(model instanceof Connection)) {
-    model = new Connection(model);
-  }
-  var url = model.driver_url;
-  var opts = model.driver_options;
-  var call;
-
-  var onConnected = function(err, db) {
-    if (err) {
-      debug('%s connection attempts to %s failed.  giving up.', url, call.getNumRetries());
-      return done(err);
-    }
-    debug('Successfully connected to %s after %s attempts!', url, call.getNumRetries());
-    done(null, db);
-  };
-
-  call = backoff.call(MongoClient.connect, url, opts, onConnected);
-  call.setStrategy(new backoff.ExponentialStrategy({
-    randomisationFactor: 0,
-    initialDelay: 500,
-    maxDelay: 10000
-  }));
-
-  call.on('backoff', function(number, delay) {
-    debug('connect attempt #%s failed.  retrying in %sms...', number, delay);
-  });
-  call.failAfter(25);
-  call.start();
-}
 
 function connect(model, done) {
   if (!(model instanceof Connection)) {
@@ -128,16 +97,10 @@ function connect(model, done) {
   });
 }
 
-if (process.env.MONGODB_BACKOFF) {
-  exports = connectWithBackoff;
-} else {
-  exports = connect;
 }
 
 exports.loadOptions = loadOptions;
 exports.validateURL = validateURL;
-exports.connectWithBackoff = connectWithBackoff;
 exports.connect = connect;
-
 
 module.exports = exports;
